refactor: move steps to separate file
